requirements.txt
google-cloud-storage
ffmpeg-python
opencv-python 